+++ modules/object.mjs
@@ -1,16 +1,20 @@
 import { ObjectManager } from "@filebase/sdk";
+import inquirer from "inquirer";
+import Table from "tty-table";
 
 export default class ObjectModule {
   constructor(program, credentials) {
-    program
-      .command("object upload <key> <source>")
+    const subcommand = program.command("object");
+
+    subcommand
+      .command("upload <key> <source>")
       .option("-m, --metadata <metadata>")
       .option("-b, --bucket <bucket>")
       .description("creates a new name with the specified label")
       .action(async (label, cid, options) => {
         const objectManager = new ObjectManager(
-          credentials.get("key"),
-          credentials.get("secret"),
+          await credentials.get("key"),
+          await credentials.get("secret"),
         );
         //todo: fixup to support file and directory uploads based on the source
         let nameOptions = {};
@@ -20,65 +24,90 @@ export default class ObjectModule {
         await objectManager.upload(label, cid, nameOptions);
       });
 
-    program
-      .command("object get <key>")
+    subcommand
+      .command("get <key>")
       .option("-b, --bucket <bucket>")
       .description("gets a object with the specified key")
       .action(async (key) => {
         const objectManager = new ObjectManager(
-          credentials.get("key"),
-          credentials.get("secret"),
+          await credentials.get("key"),
+          await credentials.get("secret"),
         );
         await objectManager.get(key);
       });
 
-    program
-      .command("object download <key>")
+    subcommand
+      .command("download <key>")
       .option("-b, --bucket <bucket>")
       .option("-d, --destination <destination>")
       .description("downloads a object with the specified key")
       .action(async (key) => {
         const objectManager = new ObjectManager(
-          credentials.get("key"),
-          credentials.get("secret"),
+          await credentials.get("key"),
+          await credentials.get("secret"),
         );
         await objectManager.download(key);
       });
 
-    program
-      .command("object delete <key>")
+    subcommand
+      .command("delete <key>")
       .option("-b, --bucket <bucket>")
       .description("deletes a name with the specified label")
       .action(async (key) => {
         const objectManager = new ObjectManager(
-          credentials.get("key"),
-          credentials.get("secret"),
+          await credentials.get("key"),
+          await credentials.get("secret"),
         );
-        //TODO: Confirm with inquirer
-        await objectManager.delete(key);
+        const answers = await inquirer.prompt([
+          {
+            type: "input",
+            name: "confirm_delete",
+            message: `Are you sure you want to delete the object with key [${key}]? Y/n`,
+          },
+        ]);
+        if (answers["confirm_delete"] === "Y") {
+          await objectManager.delete(key);
+          console.log(`Deleted Object: ${key}`);
+        }
       });
 
-    program
-      .command("object list")
+    subcommand
+      .command("list")
       .option("-b, --bucket <bucket>")
       .description("lists the objects")
       .action(async () => {
         const objectManager = new ObjectManager(
-          credentials.get("key"),
-          credentials.get("secret"),
+          await credentials.get("key"),
+          await credentials.get("secret"),
         );
-        await objectManager.list();
+        const objects = (await objectManager.list()).Contents;
+        const table = Table(
+          [
+            { value: "Key", alias: "name" },
+            { value: "CID", alias: "cid" },
+            { value: "Size", alias: "size" },
+            { value: "LastModified", alias: "last_modified" },
+          ],
+          objects,
+          undefined,
+          {
+            borderStyle: "solid",
+            borderColor: "white",
+            truncate: true,
+          },
+        ).render();
+        console.log(table);
       });
 
-    program
-      .command("object copy <key> <destinationBucket>")
+    subcommand
+      .command("copy <key> <destinationBucket>")
       .option("-b, --bucket <bucket>")
       .option("-k, --key <destinationKey>")
       .description("copies the objects")
       .action(async (key, destinationBucket, options) => {
         const objectManager = new ObjectManager(
-          credentials.get("key"),
-          credentials.get("secret"),
+          await credentials.get("key"),
+          await credentials.get("secret"),
         );
         let objectCopyOptions = {};
         await objectManager.copy(key, destinationBucket, objectCopyOptions);
